add virtualbox install

+++ ubuntu/precise-12.04/workstation/install_ws_ubuntu_12.04.sh
@@ -9,6 +9,10 @@ wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-ke
 sudo cp assets/etc/apt/sources.list.d/opera.list /etc/apt/sources.list.d/
 wget -q http://deb.opera.com/archive.key -O- | sudo apt-key add -
 
+#virtualbox
+sudo cp assets/etc/apt/sources.list.d/virtualbox.list /etc/apt/sources.list.d/
+sudo apt-key adv --keyserver keyserver.ubuntu.com --recv-keys 98AB5139
+
 #medibuntu
 sudo cp assets/etc/apt/sources.list.d/medibuntu.list /etc/apt/sources.list.d/
 wget -q http://packages.medibuntu.org/medibuntu-key.gpg -O- | sudo apt-key add -
@@ -36,6 +40,7 @@ sudo apt-get -y install libreoffice graphviz gimp gnucash keepassx keepass2 keep
 sudo apt-get -y install clamav clamav-freshclam
 sudo apt-get -y install gnome-shell gnome-panel ubuntu-tweak grub-customizer compizconfig-settings-manager compiz-plugins-extra ubuntu-restricted-extras
 sudo apt-get -y install opera google-chrome-stable
+sudo apt-get -y install virtualbox-4.1
 
 # requirements for official skype 4
 sudo apt-get -y install libqt4-core libqt4-dbus
@@ -43,10 +48,19 @@ sudo apt-get -y install libqt4-core libqt4-dbus
 # requirements for mysql workbench
 sudo apt-get -y install mysql-common python-paramiko python-pysqlite2 libctemplate0 libgtkmm-2.4-1c2a
 
-wget https://launchpad.net/ubuntu/+source/libzip/0.9.3-1/+build/1728114/+files/libzip1_0.9.3-1_amd64.deb
-sudo dpkg -i libzip1_0.9.3-1_amd64.deb
-rm libzip1_0.9.3-1_amd64.deb
+hasPkg1=`dpkg -l | grep -c 'libzip1'`
+if [ $hasPkg1 == 1]; 
+then
+    wget https://launchpad.net/ubuntu/+source/libzip/0.9.3-1/+build/1728114/+files/libzip1_0.9.3-1_amd64.deb
+    sudo dpkg -i libzip1_0.9.3-1_amd64.deb
+    rm libzip1_0.9.3-1_amd64.deb
+fi
 
-wget http://launchpadlibrarian.net/94808408/libmysqlclient16_5.1.58-1ubuntu5_amd64.deb
-sudo dpkg -i libmysqlclient16_5.1.58-1ubuntu5_amd64.deb
-rm libmysqlclient16_5.1.58-1ubuntu5_amd64.deb
+hasPkg2=`dpkg -l | grep -c 'libmysqlclient16'`
+if [ $hasPkg2 == 1]; 
+then
+    wget http://launchpadlibrarian.net/94808408/libmysqlclient16_5.1.58-1ubuntu5_amd64.deb
+    sudo dpkg -i libmysqlclient16_5.1.58-1ubuntu5_amd64.deb
+    rm libmysqlclient16_5.1.58-1ubuntu5_amd64.deb
+fi
+
